feat: Deprecate `PROPOSAL_TIME_GAP` (#50)
* contracts: accept l1Head from other proposals

* fpvm: prevent zero output proofs

* contracts: remove PROPOSAL_TIME_GAP

* update documentation

* refactor proposer/validator

* contracts: revert to proveTrailFault

* validator: revert to trail faults

* fail sync_proposal if op-node is behind

* independetly delayed proposal processing

* tune down op sync concurrency to 128 tasks

* special exit code on insufficient l1 data

* track all l1 heads

* advance proving l1 head on insufficient data error

* test coverage

* reenable validator persistence

* test premature l1 head adjustments

* prune old l1 head data

* fix

* prioritize proposal on startup

* patch devnet validator

* controlled safe l2 head delays + better output cache use

* controlled l1 head delays

* trigger pruning on resolution updates

* logging nits

* fix flaky test

* 1.0.0-rc.1

+++ crates/contracts/foundry/src/KailuaLib.sol
@@ -63,7 +63,7 @@ error NotProposed();
 error NoConflict();
 
 // 0x9276ab5a
-/// @notice Thrown when proposing before the proposal gap timeout
+/// @notice Thrown when proposing before the minimum creation time
 error ProposalGapRemaining(uint256 currentTime, uint256 minCreationTime);
 
 // 0x1434391f
@@ -140,7 +140,7 @@ library KailuaKZGLib {
     uint256 internal constant FIELD_ELEMENTS_PER_BLOB_PO2 = 12;
 
     /// @notice The number of field elements in a single blob
-    uint256 internal constant FIELD_ELEMENTS_PER_BLOB = (1 << FIELD_ELEMENTS_PER_BLOB_PO2);
+    uint256 internal constant FIELD_ELEMENTS_PER_BLOB = uint64(1 << FIELD_ELEMENTS_PER_BLOB_PO2);
 
     /// @notice The index of the blob containing the FE at the provided offset
     function blobIndex(uint256 outputOffset) internal pure returns (uint256 index) {
